Convert to AssetExtension
