--- src/loaders/index.ts
@@ -1,2 +1,0 @@
-export * from './detectSound';
-export * from './loadSound';
